Dia 3 - Integración de Aplicaciones y Protocolos SSO/Ejemplos/scripts/usuarios/requirements.txt
python-keycloak==4.0.0
requests
